requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
psycopg2=2.9.3
telebot-calendar=1.2
python-telegram-bot-calendar=1.0.5